# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/schedules": {
      "get": {
        "operationId": "SchedulesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ScheduleDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List schedules (filter by status/type/target/date/priority/search).",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "SchedulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a schedule (with optional targets).",
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/conflicts": {
      "get": {
        "operationId": "SchedulesController_conflicts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "All overlapping ACTIVE schedule pairs (shared screens + winner).",
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}": {
      "get": {
        "operationId": "SchedulesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule detail.",
        "tags": [
          "schedules"
        ]
      },
      "patch": {
        "operationId": "SchedulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a schedule.",
        "tags": [
          "schedules"
        ]
      },
      "delete": {
        "operationId": "SchedulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a schedule.",
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}/validate": {
      "get": {
        "operationId": "SchedulesController_validate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validation summary (playlist/targets, orientation + conflict warnings).",
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}/pause": {
      "post": {
        "operationId": "SchedulesController_pause",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}/resume": {
      "post": {
        "operationId": "SchedulesController_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}/archive": {
      "post": {
        "operationId": "SchedulesController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}/targets": {
      "post": {
        "operationId": "SchedulesController_addTarget",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddScheduleTargetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a target (SCREEN/SCREEN_GROUP/LOCATION/COMPANY).",
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{id}/targets/{targetId}": {
      "delete": {
        "operationId": "SchedulesController_removeTarget",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a target.",
        "tags": [
          "schedules"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wizer Signage API",
    "description": "Wizer Signage — multi-tenant digital signage SaaS platform REST API.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 137,
            "description": "Total matching rows, ignoring pagination."
          },
          "totalPages": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "DeletedResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "PlaylistRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "status"
        ]
      },
      "ScheduleTargetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "SCREEN",
              "SCREEN_GROUP",
              "LOCATION",
              "COMPANY"
            ]
          },
          "targetId": {
            "type": "string",
            "description": "The referenced entity, or 'company' for a company-wide target."
          }
        },
        "required": [
          "id",
          "targetType",
          "targetId"
        ]
      },
      "ScheduleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Owning tenant. Derived from the caller's token, never the body."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "playlistId": {
            "type": "object",
            "nullable": true
          },
          "playlist": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlaylistRefDto"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "ARCHIVED"
            ]
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "RECURRING",
              "ALWAYS"
            ]
          },
          "priority": {
            "type": "number",
            "description": "Higher wins when two schedules cover the same screen and time."
          },
          "startDate": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "endDate": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "startTime": {
            "type": "object",
            "nullable": true,
            "example": "09:00"
          },
          "endTime": {
            "type": "object",
            "nullable": true,
            "example": "17:30"
          },
          "isAllDay": {
            "type": "boolean"
          },
          "daysOfWeek": {
            "description": "0 = Sunday.",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "timezone": {
            "type": "object",
            "nullable": true,
            "example": "Asia/Riyadh",
            "description": "IANA zone the times resolve against; a bare city name runs the schedule hours off."
          },
          "recurrence": {
            "type": "object",
            "nullable": true
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "archivedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleTargetDto"
            }
          },
          "targetCount": {
            "type": "number",
            "description": "targets.length, computed — not a column."
          }
        },
        "required": [
          "id",
          "companyId",
          "createdAt",
          "updatedAt",
          "name",
          "status",
          "scheduleType",
          "priority",
          "isAllDay",
          "daysOfWeek",
          "targets",
          "targetCount"
        ]
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "AddScheduleTargetDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}